+++ neural-networks/object-tracking/deepsort-tracking/utils/deepsort_tracking.py
@@ -1,78 +1,11 @@
-from dataclasses import dataclass
-from enum import Enum
-
-import cv2
 import depthai as dai
 from deep_sort_realtime.deepsort_tracker import DeepSort
 
 from .detected_recognitions import DetectedRecognitions
-from .labels import LABELS
-
-
-@dataclass
-class ColoredLabel:
-    label: str
-    color: tuple[int, int, int]
-
-
-class TextPosition(Enum):
-    TOP_LEFT = 0
-    MID_LEFT = 1
-    BOTTOM_LEFT = 2
-    TOP_MID = 10
-    MID_MID = 11
-    BOTTOM_MID = 12
-    TOP_RIGHT = 20
-    MID_RIGHT = 21
-    BOTTOM_RIGHT = 22
-
-
-class VisualizedTracklets(dai.Tracklets):
-    def getVisualizationMessage(self):
-        img_annotations = dai.ImgAnnotations()
-        img_annotations.setTimestamp(self.getTimestamp())
-        img_annotations.setSequenceNum(self.getSequenceNum())
-        img_annotation = dai.ImgAnnotation()
-        for tracklet in self.tracklets:
-            if tracklet.status == dai.Tracklet.TrackingStatus.TRACKED:
-                pts_annot = dai.PointsAnnotation()
-                pts_annot.outlineColor = dai.Color(0.0, 1.0, 0.0)
-                pts_annot.points.append(dai.Point2f(tracklet.roi.x, tracklet.roi.y))
-                pts_annot.points.append(
-                    dai.Point2f(tracklet.roi.x + tracklet.roi.width, tracklet.roi.y)
-                )
-                pts_annot.points.append(
-                    dai.Point2f(
-                        tracklet.roi.x + tracklet.roi.width,
-                        tracklet.roi.y + tracklet.roi.height,
-                    )
-                )
-                pts_annot.points.append(
-                    dai.Point2f(tracklet.roi.x, tracklet.roi.y + tracklet.roi.height)
-                )
-                pts_annot.type = dai.PointsAnnotationType.LINE_LOOP
-                pts_annot.thickness = 2
-                img_annotation.points.append(pts_annot)
-                txt_annot = dai.TextAnnotation()
-                txt_annot.fontSize = 25
-                txt_annot.text = f"{LABELS[tracklet.srcImgDetection.label]} {round(tracklet.srcImgDetection.confidence * 100):<3}% ID: {tracklet.id}"
-                txt_annot.position = dai.Point2f(tracklet.roi.x, tracklet.roi.y)
-                txt_annot.textColor = dai.Color(0.0, 0.0, 0.0)
-                txt_annot.backgroundColor = dai.Color(0.0, 1.0, 0.0)
-                img_annotation.texts.append(txt_annot)
-
-        img_annotations.annotations.append(img_annotation)
-        return img_annotations
+from .visualized_tracklets import VisualizedTracklets
 
 
 class DeepsortTracking(dai.node.HostNode):
-    ALPHA = 0.15
-    FONT_FACE = 0
-    LINE_TYPE = cv2.LINE_AA
-    FONT_SHADOW_COLOR = (0, 0, 0)
-    FONT_COLOR = (255, 255, 255)
-    TEXT_PADDING = 10
-
     def __init__(self) -> None:
         super().__init__()
         self._tracker = DeepSort(
@@ -99,10 +32,11 @@ class DeepsortTracking(dai.node.HostNode):
     def process(
         self, img_frame: dai.ImgFrame, detected_recognitions: dai.Buffer
     ) -> None:
+        assert isinstance(detected_recognitions, DetectedRecognitions)
+
         frame = img_frame.getCvFrame()
         img_height, img_width, _ = frame.shape
 
-        assert isinstance(detected_recognitions, DetectedRecognitions)
         detections = detected_recognitions.img_detections.detections
         recognitions = detected_recognitions.nn_data
