+++ neural-networks/object-tracking/deepsort-tracking/utils/visualized_tracklets.py
@@ -1,0 +1,41 @@
+import depthai as dai
+
+from .labels import LABELS
+
+
+class VisualizedTracklets(dai.Tracklets):
+    def getVisualizationMessage(self):
+        img_annotations = dai.ImgAnnotations()
+        img_annotations.setTimestamp(self.getTimestamp())
+        img_annotations.setSequenceNum(self.getSequenceNum())
+        img_annotation = dai.ImgAnnotation()
+        for tracklet in self.tracklets:
+            if tracklet.status == dai.Tracklet.TrackingStatus.TRACKED:
+                pts_annot = dai.PointsAnnotation()
+                pts_annot.outlineColor = dai.Color(0.0, 1.0, 0.0)
+                pts_annot.points.append(dai.Point2f(tracklet.roi.x, tracklet.roi.y))
+                pts_annot.points.append(
+                    dai.Point2f(tracklet.roi.x + tracklet.roi.width, tracklet.roi.y)
+                )
+                pts_annot.points.append(
+                    dai.Point2f(
+                        tracklet.roi.x + tracklet.roi.width,
+                        tracklet.roi.y + tracklet.roi.height,
+                    )
+                )
+                pts_annot.points.append(
+                    dai.Point2f(tracklet.roi.x, tracklet.roi.y + tracklet.roi.height)
+                )
+                pts_annot.type = dai.PointsAnnotationType.LINE_LOOP
+                pts_annot.thickness = 2
+                img_annotation.points.append(pts_annot)
+                txt_annot = dai.TextAnnotation()
+                txt_annot.fontSize = 25
+                txt_annot.text = f"{LABELS[tracklet.srcImgDetection.label]} {round(tracklet.srcImgDetection.confidence * 100):<3}% ID: {tracklet.id}"
+                txt_annot.position = dai.Point2f(tracklet.roi.x, tracklet.roi.y)
+                txt_annot.textColor = dai.Color(0.0, 0.0, 0.0)
+                txt_annot.backgroundColor = dai.Color(0.0, 1.0, 0.0)
+                img_annotation.texts.append(txt_annot)
+
+        img_annotations.annotations.append(img_annotation)
+        return img_annotations
